requirements.txt
pandas=="1.3.3"
fastparquet=="0.7.1"
opencv-python=="4.5.3.56"
google-cloud-vision=="2.4.4"
tensorflow=="2.6.0"
tensorflow-gpu=="2.6.0"
numpy=="1.19.5"
pybind11=="2.8.0"
fastwer=="0.1.3"
luigi=="3.0.3"
tqdm=="4.62.3"
matplotlib=="3.4.3"